Change bounding box types from mat4 to dmat4 for double precision support
Update all bounding box related code to use dmat4/std::optional<dmat4>
instead of mat4/std::optional<mat4> to align with the larger PR that
changed SpatialEntity and Camera to use double precision.

Changes:
- zoomoptions.h: boundingBox std::function now returns optional<dmat4>
- boundingbox.h/cpp: all boundingBox and minExtentBoundingBox functions
  return dmat4; internal vec3/float changed to dvec3/double
- camerautils.h/cpp: boundingBox parameters changed to const dmat4&;
  internal computations updated to double precision
- cameratools.h/cpp: calculateFovBounds and perspectiveZoom updated for dmat4;
  camPts/finalPts arrays changed to dvec3
- cameraproperty.h/cpp: getBoundingBox_ member and constructor parameters
  changed to std::function<std::optional<dmat4>()>
- plottinggl axis helpers: getBoundingBox_ and related functions use dmat4;
  axisprocessorhelper calcScaleFactor functions updated to dmat4
- basegl cameracomponent/periodicitygl/instancerenderer: bounding box
  types changed to dmat4
- oit module: all Rasterizer::boundingBox() virtual functions return
  optional<dmat4>; transformrasterization uses dmat4 cast
- tetramesh: boundingBox lambda wraps mat4 result in dmat4 cast
- base/meshcameraalgorithms: use dvec3/dmat4 for computeNearFar call
- python3 bindings: setCameraView overloads updated to dmat4
- boundingbox unit tests: updated to use dmat4 and double literals

+++ include/inviwo/core/algorithm/camerautils.h
@@ -91,7 +91,7 @@ IVW_CORE_API float fovxDegreesFrom(float fovyDegrees, float aspect);
  * @param updateLookRanges   the camera's look-to/look-from ranges are updated if Yes @see
  * setCameraLookRanges
  */
-IVW_CORE_API void setCameraView(CameraProperty& cam, const glm::mat4& boundingBox,
+IVW_CORE_API void setCameraView(CameraProperty& cam, const glm::dmat4& boundingBox,
                                 glm::dmat3 view, float fitRatio = 1.05f,
                                 UpdateNearFar updateNearFar = UpdateNearFar::No,
                                 UpdateLookRanges updateLookRanges = UpdateLookRanges::No,
@@ -113,7 +113,7 @@ IVW_CORE_API void setCameraView(CameraProperty& cam, const glm::mat4& boundingBo
  * @param updateLookRanges   the camera's look-to/look-from ranges are updated if Yes @see
  * setCameraLookRanges
  */
-IVW_CORE_API void setCameraView(CameraProperty& cam, const glm::mat4& boundingBox,
+IVW_CORE_API void setCameraView(CameraProperty& cam, const glm::dmat4& boundingBox,
                                 float fitRatio = 1.05f,
                                 UpdateNearFar updateNearFar = UpdateNearFar::No,
                                 UpdateLookRanges updateLookRanges = UpdateLookRanges::No,
@@ -136,7 +136,7 @@ IVW_CORE_API void setCameraView(CameraProperty& cam, const glm::mat4& boundingBo
  * @param updateLookRanges   the camera's look-to/look-from ranges are updated if Yes @see
  * setCameraLookRanges
  */
-IVW_CORE_API void setCameraView(CameraProperty& cam, const glm::mat4& boundingBox, Side side,
+IVW_CORE_API void setCameraView(CameraProperty& cam, const glm::dmat4& boundingBox, Side side,
                                 float fitRatio = 1.05f,
                                 UpdateNearFar updateNearFar = UpdateNearFar::No,
                                 UpdateLookRanges updateLookRanges = UpdateLookRanges::No,
@@ -154,7 +154,7 @@ IVW_CORE_API void setCameraView(CameraProperty& cam, const glm::mat4& boundingBo
  * @param maxZoomFactor determines how far away from the volume the user will be able to zoom
  * out.
  */
-IVW_CORE_API void setCameraLookRanges(CameraProperty& cam, const glm::mat4& boundingBox,
+IVW_CORE_API void setCameraLookRanges(CameraProperty& cam, const glm::dmat4& boundingBox,
                                       float maxZoomFactor = defaultZoomFactor);
 
 /**
@@ -163,7 +163,7 @@ IVW_CORE_API void setCameraLookRanges(CameraProperty& cam, const glm::mat4& boun
  * factor.
  * @see setCameraLookRanges
  */
-IVW_CORE_API std::pair<float, float> computeCameraNearFar(const glm::mat4& boundingBox,
+IVW_CORE_API std::pair<float, float> computeCameraNearFar(const glm::dmat4& boundingBox,
                                                           float maxZoomFactor = defaultZoomFactor,
                                                           float farNearRatio = defaultFarNearRatio);
 
@@ -174,7 +174,7 @@ IVW_CORE_API std::pair<float, float> computeCameraNearFar(const glm::mat4& bound
  * @see computeCameraNearFar
  * @see setCameraLookRanges
  */
-IVW_CORE_API void setCameraNearFar(CameraProperty& cam, const glm::mat4& boundingBox,
+IVW_CORE_API void setCameraNearFar(CameraProperty& cam, const glm::dmat4& boundingBox,
                                    float maxZoomFactor = defaultZoomFactor,
                                    float farNearRatio = defaultFarNearRatio);
